init: Add reboot fs integrity test
This CL introduces a new test to verify the file system integrity after
a reboot, specifically checking for issues related to unmount
operations. Currently, the test checks whether UMOUNT_DATA_FAIL_MESSAGE
(defined in reboot.h) appears in the pstore log.

The test first checks for the existence of the pstore console logs and
skips the test if they are not available.

If the pstore is available, it reads logs from the pstore console file
and checks for the presence of specific log messages related to unmount
failures.

Bug: 407152708
Test: atest RebootFsIntegrityTest passed
Test: Revert ag/32872683, atest RebootFsIntegrityTest failed

Change-Id: I99f93573169dfdb15128d18e11c417961554b2c3

Former-commit-id: 2080177d4da2f01895b14537263aac5378c11965