init: reboot: Remove `reboot_target` arg from RebootMonitorThread()
The `reboot_target` is only relevant for reboot, not shutdown. The
RebootMonitorThread() never directly reboots--whenever it needs to
reboot it writes "c" to `/proc/sysrq`. That means we don't need the
`reboot_target`. Remove it.

NOTE: if you dig into RebootSystem() it looks as if we need to keep
`reboot_target` in the case of `ANDROID_RB_THERMOFF`. However, if
you follow through the code you can see that in HandlePowerctlMessage()
whenever we set `ANDROID_RB_THERMOFF` we always leave `reboot_target`
as "".

Bug: 409835922
Test: Hack a timeout in and see shutdown/reboot still work.
Change-Id: I9390bc30ae0fb555ee018a3f3c2697d379b6c881

Former-commit-id: 1695b29df02b1d9a24693963e07df6e72e5211f5