keymint_in_vm: Use ICommService for VM Communication
This cl refactors the KeyMint HALs to communicate with the Trusty
VM via the RPC binder service ICommService running within the VM
instead of a direct TIPC vsock connection.

Previously, the HAL was responsible for creating and managing the
vsock connection to the VM. In the new model, the early_virtmgr
of AVF handle the vsock setup.

A key benefit of this refactoring is the complete decoupling from
the Trusty driver stack. The HAL connecting to the Trusty VM now
communicates using only the generic `libbinder_rs` and no longer
requires any Trusty-specific libraries.

Bug: 404289721
Test: launch KeyMint connecting to trusty VM
Change-Id: Ic449b85f7cf4a3feaa16c64414c8b30372b9f5cc

Former-commit-id: c8ec59f28adce5127814dcc15ac4be5e155cf69c