Add Android.bp
We want to export our check_for_ufs script so it can be pulled in
by automated test suites in other parts of the tree.

We make a separate build file to avoid cluttering the Android.bp
in our parent directory, and also to allow test suites to directly
reference this script by name, without needing the "test" directory.

Bug: 409357709
Test: Build with a (WIP) test suite CL was able to use this
Flag: EXEMPT for test infra change
Change-Id: I3ca82a9ae1126db8907f174bfbda829432d7bbb5

Former-commit-id: 5d4d430e98d6efbd5189f6bdb49cde3ce8e0509c