check_for_ufs: Normalize output
This CL does not change any of the UFS vs. eMMC detection logic.

We plan on using the output of this script in automated test
suites.  So we look to simplify with just a single line of
output through expected paths.  We use a single exit function
to normalize this.

Bug: 409357709
Test: Run through on a device with eMMC and one with UFS
Flag: EXEMPT for testing script
Change-Id: Ide2ece12fe54770b1bad3af31c221d63d0241dee

Former-commit-id: b8ea59c285cf9d0130ae8036118a025d0ee57bc3