fs_mgr: skip /data check when mapping
ImageManager::MapImageDevice() checks if /data is backed by
device-mapper. The check fails when the method is invoked before /data
is mounted, which is what apexd does with the mount_before_data flag.

The check can be skipped if the image header is not found.

Bug: 385834023
Flag: com.android.apex.flags.mount_before_data
Test: adb install <apex> && adb reboot
Change-Id: I841d956a5c2de103f4e84894103e0560cc96922f

Former-commit-id: 4563d5dc22ef97952221e2c0a04e88c5f6de95bb