Clean up some debuggerd test data.
The code uses tmpfile to generate the security policy file for
seccomp tests. Switch this to use the libbase TemporaryFile since
it's easier to use.

Also, move all usage of TemporaryDir into the main test code
body and out of StartProcess. Otherwise, the directory will not
be cleaned up properly.

Test: All unit tests pass and no leftover files in /data/local/tmp.
Change-Id: I6648befd9fd75a703a25b5be25e69a9e021b4b4e

Former-commit-id: 1ed8acb7be9381fbbe801cd185de2ee63951b136