check_for_ufs.sh: Remove need for root
Some of our use cases for this scripts aren't assured to have
root access on the device.  So we revamp this script with a new
approach which doesn't appear to need root on the devices we've
tested it on.

This has the advantage of reducing the runtime of the script on
UFS devices (no longer need to run a multi-second "find" command).

We also expand our use of 'local' for clarity, and make the error
from our command detection logic more directly actionable.

Bug: 419359357
Test: Run without root on a device with UFS and one with eMMC
Flag: EXEMPT for a bash script that is not used in the runtime
Change-Id: I414f8a446a3a2a7b8c72dba15da948539f00301d

Former-commit-id: c461fc9c9fe6664edb939e56f68191c0541b3473