Clean up debuggerd tests.
Add a kTimeout response time to detect a real timeout and make errors
more clear.

Modify the tests to modify the FinishIntercept function so that it
asserts when there is a failure rather than having to always pass
in a value and add an extra assert in every single test.

Test: All unit tests pass.
Change-Id: I0251fe15ee877f7129b878c52f65a017d6deefcd

Former-commit-id: 4ea412ee763aa484bec193e3140aa1dbed9a6db9