fastbootd: Remove an erroneous partition size check.
It is possible that when flashing a file that is not 4KB aligned, the
sparse size will be larger than the unsparse size. The code incorrectly
validates the sparse size, rather then the underlying data size.

Simply remove this check as it's redundant (lseek/write will fail if out
of bounds).

Bug: 415214175
Test: manual test
Change-Id: Iff7ed66925e05aa0ab8727755bcd4f5083126a06

Former-commit-id: 6bf97d49e1d4c5276abf90582a4ecb4204286a2f