fastboot: Fix issues flashing unaligned files.
If you try to flash an unaligned file, and the file requires
sparsification, libsparse will either infinite loop or error out.

As a workaround for this, detect if we can use a 512-byte sparse chunk
size instead of 4096. If not, then we explicitly fail.

In order to add a test, this CL also moves more functions into the
IFastBootDriver interface.

Bug: 415214175
Test: fastboot_test
Change-Id: I46e350498302dff6071af55b27a6009a607bfa6a

Former-commit-id: 0f7a09d077faf0fa21e585c3d588c54ea6cd9823