init: Move Coldboot to new files
As we plan to provide an option to introduce thread pool to coldboot,
separate the coldboot code and move it to coldboot.cpp and
coldboot_subprocesses.cpp and make it easier to switch between the
original implementation and the new thread pool implementation based on
a feature flag, which is going to be introduced in future CLs. The
actual code and behavior remains the same.

Bug: 421025789
Test: build passes and DUT boots
Change-Id: Id2cd7d83049c681a92fb3602bcc5886688f9f35c

Former-commit-id: 6de8916af7676f5ff30fc475f3b38987fb6abf67