Add Priority to the ThreadPool
This commit allows to queue tasks with different priorities. Each Task
is tied with a priority represented by an integer (the smaller the more
prioritized).
In ueventd coldboot, the priorities are listed in an enum
`ThreadPoolPriority`. ModaliasHandler's tasks have the highest priority
as kernel module loading has dependencies, restorecon has the second
highest, and the rest has the lowest.

Bug: 421025789
Test: DUT boots
Change-Id: I03536e9c642c30d61ff686c9988f2c530ad7cdb0

Former-commit-id: f9176a2806708a1f449b66655af892cbc187b4f3