Stop assuming getpid() is broken.
We're removing the pid caching from bionic, so we don't need to work around it here, nor should a test assume getpid() is broken.

What to do about the equivalent gettid() problem is still undecided, so leave that for now. (But realistically, we should probably remove that too, like glibc did.)

Change-Id: I64168a8d959a81894dcaa3a98ae0bc5bba9e575f

Former-commit-id: 4c46aa583f3bf77db80afe4ea482a4a6a94dcb1c