CtsInittestCases: fix enter_namespace_net_preserves_mounts test
One of the rootfs mounts, /sys/kernel/debug/tracing, is not mounted by
init, but by the kernel on first access to its contents.

As a result, in case /sys/kernel/debug/tracing *is* mounted in the root
namespace at the point of executing the test, the test relies on
_something_ in the newly created network namespace accessing
/sys/kernel/debug/tracing. This causes the test to fail with some system
configurations.

This CL makes the test explicitly trigger kernel to automount the
directory.

Bug: 425853301
Test: 1. adb shell mount -t debugfs debugfs /sys/kernel/debug
Test: 2. adb shell touch /sys/kernel/debug/tracing/trace_marker
Test: 3. atest CtsInitTestCases
Change-Id: Ieee7c248db9580e531440a29cbadef9656bbea44

Former-commit-id: b6dfac134c677b63801d31780cd2c1b1114e7095