Thaw any frozen processes when zygote restarts
If system_server crashes, any processes that were frozen remain
frozen.  system_server might not recognize this when it restarts.
This change thaws all frozen processes before restarting zygote (which
always restarts when system_server must be restarted).

Spot tests show that the time spent thawing processes is on the order
of 100ms.  Note that this extra time is only incurred when zygote must
be restarted.  It is not incurred on first boot.

Tested with script that killed system_server repeatedly, verifying
that any processes frozen before the kill were not frozen after
system_server was restarted.

Flag: EXEMPT bug-fix
Bug: 414581596
Test: presubmit

Change-Id: Ib34da13b4e837bd1bdac7a7e541e6763c8c27c79

Former-commit-id: 0800039fba104113ea873c9d5d14291ddbe254d8