init: leave perms on /proc/cmdline at 0444
Relax the permissions of /proc/cmdline to the default of 0444. Access
is still restricted via SELinux policies. This change allows us to allow
adbd access to /proc/cmdline, which is needed to support CTS testing
which validates the kernel command line. /proc/cmdline may contain
unique device identifiers, and we don't want to open it up to everyone.

Bug: 437077186
Test: boot to home
Test: Run modified Cts KernelConfigTest (b/384977724)
Change-Id: Ia3b47085aa89a4dec272d74ce5ffca0f8bcad7fc

Former-commit-id: 03fb93ca490beaecd5f6d0c9ef446c1c109aaae6