Update bootstats to use ms for boot_complete
This change updates bootstats to use milliseconds instead of seconds for
the boot_complete metrics. This aligns the units with the
BootTimeEventElapsedTime atom and provides finer measurements.

Currently, there are no other consumers of this atom. Lab tests that
monitor bootstats directly may identify this CL as a culprit of a boot
time regression, but this change doesn't actually affect boot time.

Test: build, boot, check boot_complete files' mtime
Bug: 427910232
Flag: EXEMPT bugfix
Change-Id: I60b9a88697955a1a795b28716a14940a55e1c40b

Former-commit-id: a64f96356359365abab57b8cf938e93be80c8d84
Former-commit-id: 8b10d1076d265c2ffea84b6571c3b05f68d5a675